Make scenes with related list cooperate with the database
